Remove the `latest` symlink, add a 404 page
People often stumble across the URL:

https://numba.pydata.org/numba-doc/latest/

and mistake it for the latest documentation, often missing the banner
stating that the documentation is outdated.

This attempts to rectify this problem by removing the `latest` symlink
(leaving all old versioned docs intact) and adding a 404 page that
points to the latest documentation.

The intended result is that anyone attempting to view latest docs via
the old URL should follow the link on the 404 page to the current
documentation.